// src/types/home.interface.ts
export type AchievementItemType = {
  title: string
  description: string
}

export type ContactFormKey =
  | 'emptyErrorFullnameMessage'
  | 'emptyErrorEmailMessage'
  | 'emptyErrorPhoneMessage'
  | 'emptyErrorMessageMessage'
  | 'invalidErrorEmailMessage'
  | 'invalidErrorPhoneMessage'
  | 'fieldFullnameLabel'
  | 'fieldEmailLabel'
  | 'fieldPhoneLabel'
  | 'fieldMessageLabel'
  | 'fieldFullnamePlaceholder'
  | 'fieldEmailPlaceholder'
  | 'fieldPhonePlaceholder'
  | 'fieldMessagePlaceholder'